Added copy titlebar button to property "current value" window, Reworked the help documentation window to be less rubbish + added some gridlist extensions, Made properties option skip to the relevant properties (rather than landing on the element selection pane)

+++ guieditor/client/message_box.lua
@@ -44,7 +44,10 @@ function MessageBox:close()
 		self.onClose(unpack(self.onCloseArgs or {}))
 	end
 	
-	destroyElement(self.window)
+	if exists(self.window) then
+		destroyElement(self.window)
+	end
+	
 	self = nil
 end
